TextInput: Abstract common input behavior to AbstractTextInput
The AbstractTextInput class now handles all of the behavior shared
between single and multi-line text inputs. This includes everything
from cursor movement, to selection, to scrolling, to text drawing.

This abstraction was important in order to support Undo+Redo operations
easily for both input types.
